Переименовывает модули согласно новой схемы
