+++ js/picture.js
@@ -1,8 +1,23 @@
-import {createThumbnail } from './create-thumbnail.js';
-import { openPictureModal, closePictureModal } from './render-modal.js';
+import { openPictureModal, closePictureModal } from './big-picture.js';
 import { isEscKey } from './util.js';
-
+// Контейнер для миниатюр
 const thumbnailsContainer = document.querySelector('.pictures');
+
+// Нахождение шаблона миниатюры
+const thumbnailTemplate = document.querySelector('#picture').content.querySelector('.picture');
+
+// Создание узла миниатюры из объекта фотографии
+const createThumbnail = (photo) => {
+  const {url, comments, likes} = photo;
+  const thumbnail = thumbnailTemplate.cloneNode(true);
+
+
+  thumbnail.querySelector('img').src = url;
+  thumbnail.querySelector('.picture__comments').textContent = comments.length;
+  thumbnail.querySelector('.picture__likes').textContent = likes;
+
+  return thumbnail;
+};
 
 // Обработчик нажатия на клавишу ESC на оверлее просмотра большого изображения
 const onModalEscKeydown = (evt) => {
@@ -30,8 +45,8 @@ const renderThumbnail = (photo) => {
   return thumbnail;
 };
 
-const renderThumbnails = (photos) => {
+const renderPictures = (photos) => {
   thumbnailsContainer.append(...photos.map(renderThumbnail));
 };
 
-export { renderThumbnails, onModalCloseButtonClick, onModalEscKeydown };
+export { renderPictures, onModalCloseButtonClick, onModalEscKeydown };
